fix: move setContentOrder function from helpers to cmsNameMapper
